[Chore/#94] KFImage extension 파일 분리
