+++ Staccato-iOS/Staccato-iOS/Util/Extension/Ex+Image.swift
@@ -6,25 +6,9 @@
 //
 
 import SwiftUI
-import Kingfisher
 
 extension Image {
     init(_ icon: StaccatoIcon) {
         self.init(systemName: icon.rawValue)
-    }
-}
-
-extension KFImage {
-    func fillPersonImage(width: CGFloat, height: CGFloat) -> some View {
-        self.placeholder {
-            Image(.personCircleFill)
-                .resizable()
-                .foregroundStyle(.gray2)
-                .frame(width: width, height: height)
-        }
-        .resizable()
-        .aspectRatio(contentMode: .fill)
-        .frame(width: width, height: height)
-        .clipShape(Circle())
     }
 }
